Option to run specific migration(s), refs #13368
Added option, to the tools:upgrade-sql task, to run specific migrations
rather than a full upgrade. When running a specific migration the
schema version, in the database, won't be changed. The --number option
is used to specify the migration number to run and multiple migrations
can be run by separating migration numbers with commas.